skip indecipherable domain names instead of choking on them (see #2)
If the DHCP server (or other source of resolv.conf information) sends
a non-ASCII domain name or a non-ASCII search path, we have two
options -- bail on the whole thing, or ignore domain names we don't
think look reasonable and not transfer them to /etc/resolv.conf.

With this changeset, i'm taking the latter option.  We still haven't
properly resolved #2, but at least a source of DNS information which
is provides such a thing won't have the other information it offers
rejected as a result.