simplify non-resolvconf operation
When /sbin/resolvconf isn't present, and there are multiple
concurrent/interleaved sources of DNS information, there is
potentially squirrelly interaction.

In such a situation, we have three choices:

 a) handle all the sources entirely robustly (which is what
    resolvconf(8) intends to do)

 b) do something simple, explainable, and predictable, even if it is
    not particularly robust.

 c) do something complex and hard to understand.

This changeset moves resolvconf-admin's behavior in the
non-/sbin/resolvconf situation from (c) to (b), and adds documentation
to the manpage to explain the simple behavior.